[Impeller] force initialize all variants for Vulkan.
